% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starwars_util.R
\name{my_starwars_name}
\alias{my_starwars_name}
\title{Provides a random character name from Star Wars}
\usage{
my_starwars_name(female_only = FALSE)
}
\arguments{
\item{female_only}{Do you want a female character name?}
}
\value{
A single random Star Wars character name
}
\description{
Provides a random character name from Star Wars
}
